// client/src/clients/deal-system-client-definitions.ts
export class LookupRequest implements ILookupRequest {
    id!: number;
    name!: string;
    description!: string;
    active?: boolean | undefined;

    constructor(data?: ILookupRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            id: observable,
            name: observable,
            description: observable,
            active: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.active = _data["active"];
        }
    }

    static fromJS(data: any): LookupRequest {
        data = typeof data === 'object' ? data : {};
        let result = new LookupRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["active"] = this.active;
        return data; 
    }

    clone(): LookupRequest {
        const json = this.toJSON();
        let result = new LookupRequest();
        result.init(json);
        return result;
    }
}

export interface ILookupRequest {
    id: number;
    name: string;
    description: string;
    active?: boolean | undefined;
}
// ...
export enum LazyLoadedDataStateEnum {
    NotLoaded = 0,
    Loading = 1,
    Ready = 2,
}
// ...
export class LookupRequestHeader implements ILookupRequestHeader {
    currentUser?: boolean | undefined;
    id?: number | undefined;
    name!: string;
    totalCount?: number | undefined;
    currentPage?: number | undefined;
    results!: LookupRequest[];

    constructor(data?: ILookupRequestHeader) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
            if (data.results) {
                this.results = [];
                for (let i = 0; i < data.results.length; i++) {
                    let item = data.results[i];
                    this.results[i] = item && !(<any>item).toJSON ? new LookupRequest(item) : <LookupRequest>item;
                }
            }
        }
        if (!data) {
            this.results = [];
        }
        makeObservable(this, {
            currentUser: observable,
            id: observable,
            name: observable,
            totalCount: observable,
            currentPage: observable,
            results: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.currentUser = _data["currentUser"];
            this.id = _data["id"];
            this.name = _data["name"];
            this.totalCount = _data["totalCount"];
            this.currentPage = _data["currentPage"];
            if (Array.isArray(_data["results"])) {
                this.results = [] as any;
                for (let item of _data["results"])
                    this.results!.push(LookupRequest.fromJS(item));
            }
        }
    }

    static fromJS(data: any): LookupRequestHeader {
        data = typeof data === 'object' ? data : {};
        let result = new LookupRequestHeader();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["currentUser"] = this.currentUser;
        data["id"] = this.id;
        data["name"] = this.name;
        data["totalCount"] = this.totalCount;
        data["currentPage"] = this.currentPage;
        if (Array.isArray(this.results)) {
            data["results"] = [];
            for (let item of this.results)
                data["results"].push(item.toJSON());
        }
        return data; 
    }

    clone(): LookupRequestHeader {
        const json = this.toJSON();
        let result = new LookupRequestHeader();
        result.init(json);
        return result;
    }
}

export interface ILookupRequestHeader {
    currentUser?: boolean | undefined;
    id?: number | undefined;
    name: string;
    totalCount?: number | undefined;
    currentPage?: number | undefined;
    results: ILookupRequest[];
}